refactor: improve DX in rolldown_binding
