+++ crates/rolldown_binding/src/options/binding_output_options/binding_pre_rendered_chunk.rs
