--- crates/rolldown_binding/src/options/binding_output_options/types/mod.rs
@@ -1,1 +1,0 @@
-pub mod binding_advanced_chunks_options;
