+++ crates/rolldown_binding/src/types/binding_log_level.rs
@@ -1,6 +1,5 @@
-use std::fmt::{self, Display, Formatter};
-
 use napi_derive::napi;
+use std::fmt::{self, Display, Formatter};
 
 #[derive(Debug, PartialEq, Clone, Copy, Default)]
 #[napi]
