+++ crates/rolldown_binding/src/types/mod.rs
@@ -5,7 +5,6 @@ pub mod binding_normalized_options;
 pub mod binding_output_asset;
 pub mod binding_output_chunk;
 pub mod binding_outputs;
-pub mod binding_pre_rendered_chunk;
 pub mod binding_rendered_chunk;
 pub mod binding_rendered_module;
 pub mod binding_resolve_alias_item;
